% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createSlopeParameters}
\alias{createSlopeParameters}
\title{Format Slope parameters}
\usage{
createSlopeParameters(slopeshape, slopesteepness)
}
\arguments{
\item{slopeshape}{a character, one of 's-shaped', 'convex', 'concave', 'uniform'}

\item{slopesteepness}{integer percent grade of slope}
}
\value{
formatted character string input for .par file
}
\description{
Format Slope parameters
}
\examples{
createslopeParameters('s-shaped', 40)
createslopeParameters('s-shaped', 80)
createslopeParameters('uniform', 40)

}
