% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createStormFile}
\alias{createStormFile}
\title{Create Storm File}
\usage{
createStormFile(data, outfile, ...)
}
\arguments{
\item{outfile}{target .pre file.}

\item{...}{list of key:value pairs that are written to the header of the file}
}
\value{
.pre file
}
\description{
Create Storm File
}
\details{
Input dataframe should have the columns corresponding to the input .pre file, including the following columns:
\describe{
  \item{day}{day of month}
  \item{month}{month}
  \item{year}{year}
  \item{rain}{daily rainfall}
  \item{duration}{duration in hours}
  \item{Tp}{fractional time to peak intensity}
  \item{Ip}{Rainfall intensity at peak (mm / hr)}
}
}
