% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_soil_texture}
\alias{get_soil_texture}
\title{Retrieve soil texture parameters}
\usage{
get_soil_texture(texture)
}
\arguments{
\item{texture}{character}
}
\value{
named list of input parameters for given soil type
}
\description{
Retrieve soil texture parameters
}
\details{
texture may be one of "clay", "clay loam", "loam", "Loamy Sand", "sand", "sandy clay", "sandy clay loam", "sandy loam", "silt","silt loam", "silty clay", "silty clay loam"
}
\examples{
get_soil_texture('sandy loam')

}
