% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{par_defaults}
\alias{par_defaults}
\title{Default parameters}
\usage{
par_defaults()
}
\value{
A named list of default parameters used for building the RHEM .par file
}
\description{
Default parameters
}
\examples{
par_defaults()
}
