% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read.rhem}
\alias{read.rhem}
\title{Read output}
\usage{
read.rhem(f)
}
\arguments{
\item{file}{path to .out file.}
}
\description{
Read output
}
\details{
creates a temporary '.run' file and executes it
}
