% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_rhem}
\alias{run_rhem}
\title{Format climate database}
\usage{
run_rhem(parfile, prefile, executable, outfile = tempfile(), cleanup = TRUE)
}
\arguments{
\item{parfile}{path.}

\item{prefile}{path.}

\item{executable}{path to executable.}

\item{data}{A data.frame.
Run Rhem}

\item{output}{path.}
}
\description{
Format climate database
}
\details{
creates a temporary '.run' file and executes it
}
