% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{texture_df}
\alias{texture_df}
\title{Soil Texture parameters}
\usage{
texture_df()
}
\value{
A data.frame with each row corresponding to a USDA soil texture class and each column representing a parameter used for building the RHEM .par file
}
\description{
Soil Texture parameters
}
\examples{
texture_df()
}
