% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_par}
\alias{build_par}
\title{Write .par file for input to rhem}
\usage{
build_par(...)
}
\arguments{
\item{...}{named arguments. see \code{details}}
}
\value{
list of inputs for .par file
}
\description{
Write .par file for input to rhem
}
\examples{
a <- build_par() # defaults
a
unlink(a$handle)

}
